Add manifest for flatpak
Add a manifest that includes all the dependencies which are not present
in the GNOME runtime. Copy patches from the Fedora (and Debian)
packages. Include a custom patch for the libid3tag pkg-config file, as
it is not easy to include an extra source file and then install it
(without modifying the build system).